InterfaceDemo
The code implements an example of using an interface in the C# programming language. Let's see how it works in the basics:

i1 interface:

Defines an interface with two methods, display1 and display2.
Program class:

The Program class implements the i1 interface (implements i1). This means that it must provide implementations for all the methods defined in the interface.

In the Main method, the message "Hi" is printed to the console.

Implements the display1 and display2 methods required by the i1 interface. In this case, they just print messages to the console.

In the #region i1 Members part, the implementation of the interface methods is enclosed, and this is to make the code easier to read in development environments (IDEs).

In the Main method, an object of the Program class is created, and the display1 and display2 methods are called from this object.